started on getting a second view working. currently crashes
